+++ app/src/main/java/com/itg3/grp1/mobdevproject/DbTables/Users.kt
@@ -98,7 +98,7 @@ class Users(newDbHandler: SQLiteOpenHelper) : IDbTable<User>(newDbHandler)
         return success
     }
 
-    override fun update(instance: User) : Long
+    override fun update(instance: User) : Int
     {
         TODO("Not yet implemented")
     }
